Complete making all Iterator methods constexpr
Part of this was an improvement to try_for_each() which no longer requires
passing in a success value. To do this we need a TryDefault concept which
allows constructing a default success value. That allows Result<void, E> to
be Try because it is TryDefault and thus try_for_each does not need to rely
on the Try type being constructible from a success value or to be default
constructible itself (which would often make a failure value like for Option).