Define the return types of <=>
Avoid using auto so it shows up in the docs, this made us
fix the constraits on Tuple as it uncovered a mistake there. For
Tuple all the elements must be StrongOrd to be StrongOrd, else
all the elements must be Ord to be Ord. It was comparing each
element instead of all the elements.