Turn Markdown headers into named anchors that link to themselves
We formed md4c to add the functionality into the parser, so point the
git submodule to that fork. The md4c project seems to be unmaintained
which is why we did it that way. https://github.com/mity/md4c/issues/192

Then move markdown parsing to the generate step so that we can track
repeated self links per-page and increment a counter on each one to
make them unique.

Add missing Result::and_then() to support this use of Result. It still
needs tests (and the Copy overload).