Fix compilation on Clang Mac
On Mac, the size_t and uintptr_t types are not interchangeable with
uint32_t or uint64_t, so don't use them as the usize and uptr
primitive_value field types. Doing so leads to ambiguous overload
problems when calling overloads based on uint32_t and uint64_t as it
could convert to either one.

Instead, use the uint32_t or uint64_t type directly.

Include option.h in size_hint_impl.h to ensure the formattability of it
is visible always when defining the formatting of SizeHint.