Further improve/restrict const& overloads in marker types
Filed a bug upstream with Gtest about it, since it's a straightward
fix to not require these const& overloads:
https://github.com/google/googletest/issues/4350

Regardless, remove the requires clause from the marker types conversion
operators to avoid errors that are just a long list of all operator==
in the system. Only one should be used (the most matching one). Instead
put a static assert so that if the type in the marker can't be used to
construct the real type (e.g. Option from SomeMarker) then we can
display a more useful error message.

For the gtest-specific const& overloads, we need to support a different
type of conversion, from const T& to U instead of from T&& to U. So
check for that with a static assert and suggest using
EXPECT_TRUE(a == b) if the conversion would be valid outside of gtest
macros (where there's no conversion to a const lvalue). But also require
the real T&& to U conversion to be possible so that test code must be
as capable as production (just actually a bit more capable for gtest
macros due to the bug.)