Simplify into() into just holding a reference to whatever it's given
Much like SomeMarker and friends. Allow it to hold lvalue references
(like arrays, which string literals are, and like many getter functions
return). Drop ref_into() and mref_into() and move everything to into().

Found a fallible From<NonNull, T*> impl which is incorrect, From should
be infallible, and that's why Rust std::ptr::NonNull does not that that
From implementation. It could be a TryFrom impl instead though that's
not on the Rust type at the moment. So remove that From impl.