Add css and topbar for mobile
On small screens, remove the side bar and put a top bar. Remove a lot
of left/right padding and spacing to use the available space better,
and change lists that are tables into verticle blocks.

Also make the template params in a type signature a <pre> so that it
does not wrap, and we can scroll the whole signature together to the
right with a scrollbar.