Stop deleting info from type names
We were shorten type names in a hacky way but it throws away important
info like qualifiers!

So stop doing that. For fn parameters, we collect the full qualifiers
and pointer stuff, then if the type is in the database we can display
it back with all the qualifiers in a nice concise way without the raw
type-to-string text.