Avoid auto-moving from ranges and add sus::ptr::as_ref.
C++ ranges are always a "view", as they operate on pointers and not
on values, even if you have a range over an rvalue of an owning
container like vector. To move out of a range, you need to move each
individual element (oof).

When sus::iter::from_range(rvalue) happens we were assuming we can move
from the pointers in the range, but the range may not be owning, in
which case the move is invalid. So instead, iterators over ranges will
always return references, just as ranges always return pointers (or
iterators that deref to references).

In order to be able to move from a range, introduce a .moved() method on
iterators, which like copied() converts an iterator over T& to an
iterator over T. However instead of copying, it will move from each
element referenced by the iterator. Naturally this is a big footgun but
it's required to work with ranges, such as to move every element out of
a std::vector. With sus::Vec you can `.into_iter()` and work with values
directly, but not with std::vector and its range directly.

To assist in working with pointers, provide the sus::ptr::as_ref()
function that converts T* to Option<T&>. This is equivalent to the Rust
pointer::as_ref() associated function that is present on all pointer
objects.