Avoid compiler errors in Fn*Ref.
Do not put [[clang::lifetimebound]] on move contructor/assignment
parameters, as it implies to the compiler that the moved-from object
must outlive (since its is referenced by) the moved-to object.

Use contructible_from to simply the defn of from() for all Fn*Ref types.

Make the never-value-field functions noreturn so they don't need to
construct a return value or make a compiler error.