Add deduction guides for Tuple, Array, Vec. Convert Tuples
We keep the tuple() function as it stores references which is important
for making a tuple of references without spelling the whole type out.
For Vec/Array this isn't important as they don't hold references.

Make TupleMarker not copy/moveable.

Allow Tuple<X> to convert to Tuple<Y> if X converts to Y. When you're
not working with a reference in the tuple, it's fine to build concrete
Tuples and they compare them with compatible/convertible Tuples. And
static_assert() that no reference to temporary is created during the
conversion like we do in the IntoRef conversion function.

Compiler errors when you try to store IntoRef are very unclear, so add
a static assert to try make something better show up.

Update the Reverse example to use Tuple ctor type deduction.