Add round_ties and clarify the rounding behaviour of round
The round methods will match the behaviour of Rust, which matches
the C round() function except for also preserving the negative sign
on zero.

Add round_ties which will use the current rounding mode to round
ties, which by default is to the nearest even number which is
also what is specified by the IEEE 754 standard. Rust has an
unstable round_ties_even which does the same thing, but does not
need to worry about the user specifying a rounding mode, so it
can always provide a single behaviour. For this reason and to
avoid excessive calls to set the rounding mode, we provide a more
general and C++-specific version that follows the mode and does
not claim any one in particular.

We document that round_ties is much faster than round and is usually
preferable.