Add round_to_iN() methods on f32 and f64
These methods perform the same thing as `round_ties()` and then a safe
saturating cast. But that requires a separate instruction for the
rounding and the casting. Some platforms have a single instruction that
does both. So this gives access to that instruction while guarding for
UB when the value is out of bounds for the target type and saturating,
and returning 0 for NaN.