Fix compile of float rounding function tests on MSVC and test for NaNs.
I mistakenly thought fesetround() would return the previous mode, but it
returns 0 for success. So we need to call fegetround() first to be able
to restore the mode.

Changing the fp environment does not produce a reliable output in tests
so don't do it (no one changes the env anyway).