Strip `@foo.modmap` command-file from the command line
CMake 3.28 adds a modmap command file to the command line that it
writes into compile_commands.json but the file does not actually
exist outside of cmake compiling. This means the command line in
the compile_commands.json fails with an error about the missing
file:
```
error: no such file or directory: '@sus/CMakeFiles/subspace_unittests.dir/mem/addressof_unittest.cc.o.modmap'
```

To work around this we can strip out the use of the modmap command-file
for now. This will break compilation later when C++ 20 modules are being
used though. CMake will need to find a way to provide a command line in
the compile_commands.json file which can actually be used by other tools
to reproduce the compilation step.

CMake discussion: https://discourse.cmake.org/t/how-to-control-the-location-of-the-c-20-binary-module-interface-bmi-output-directory/7968/13
Fixes #437