Add support for pagefind.app searching
pagefind generates a search index from the site and inserts a
search box into the page that searches it.

Subdoc does not run pagefind automatically, because the results are
a bit subpar. Searching "option" doesn't list the Option class in the
first page of results even though the page has a 20 weight on the name
"Option". The namespace comes up first, and it has a 10 weight on the
same. Also the search box is quite big and both the box and results are
hard to configure visually as they don't provide helpful css classes to
work with.

So we should try another tool out.