Generate a search.json database for lunr searching
lunr builds an index from a set of fields, which we generate with
subdoc into 'search.json'

The webpage loads the json file and puts it into a lunr index. In
the future this could be done offline and serialized, but it would
need JS execution during Subdoc execution or during page deployment.