Fix search on Safari and handle errors and `::` queries
`::` needs to be escaped to a space, since `:` is handled specially
by lunrjs.

Errors during search get displayed on the page.

Avoid import for json files as Safari doens't support it. Instead write
a JS file which sets a variable, and load it through a script tag.

Generate search index items for things without pages of their own:
- fields
- methods
- aliases