Remove optional converstions that change the inner type
This is abnormal for C++ containers. In the future std::optional
may gain the ability to hold a reference so we don't want to convert
from references to pointers now since that would mean the wrong thing
happens later or a breaking change would be needed. Instead provide docs
saying how to go from a reference to a pointer or value.